Add custom mouse buttons to VirtualButtonCollection, change mouse controls on Net
Games can now configure what happens when the left, middle or right mouse buttons are pressed, and what happens when a touch or hold action is performed.
Net uses the left and right mouse buttons to rotate, and the middle button to lock. On touch screens the controls are unchanged.